Some comments and refactoring
