+++ app/src/main/java/com/artemiymatchin/notificationplayer/actions/LoopActionBroadcastReceiver.kt
@@ -3,7 +3,7 @@ package com.artemiymatchin.notificationplayer.actions
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import com.artemiymatchin.notificationplayer.MusicPlayer
+import com.artemiymatchin.notificationplayer.model.MusicPlayer
 
 class LoopActionBroadcastReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context?, intent: Intent?) {
