+++ app/src/main/java/com/artemiymatchin/notificationplayer/actions/PauseActionBroadcastReceiver.kt
@@ -3,8 +3,7 @@ package com.artemiymatchin.notificationplayer.actions
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import android.widget.Toast
-import com.artemiymatchin.notificationplayer.MusicPlayer
+import com.artemiymatchin.notificationplayer.model.MusicPlayer
 
 class PauseActionBroadcastReceiver : BroadcastReceiver() {
     override fun onReceive(context: Context?, intent: Intent?) {
