+++ app/src/main/res/drawable/unrepeat_icon.xml
@@ -1,4 +1,4 @@
-<vector android:height="24dp" android:tint="#0249C8"
+<vector android:height="24dp" android:tint="#FFFFFF"
     android:viewportHeight="24" android:viewportWidth="24"
     android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
     <path android:fillColor="@android:color/white"
